* Separate WIC lib definitions into separate file, and add some type hints with WIC module.
